Refactor models
 * All models have a static `fromJson` method which returns a shared
pointer to a new model created from the passed JSON.
 * All models have a `deserialize` method, which initializes the model's
values based on the information in the passed JSON object.
 * All models have a `serialize` method, which returns a JSON object
that resembles Discord's format as much as possible.
 * Pointers to the parent model are now handled by the parent model,
instead of being set in the constructor.
 * All models implicitly convert to `bool` and have an isNull method to
tell if the model is valid.
 * All models now have at least an equality and inequality overloaded
operators.
 * The order of methods and variables in the models' headers has been
changed to mirror the official Discord documentation. However, said
documentation is flawed, so full compliance will not happen.
 * Some tests have been rewritten without Qt's way of storing data, as
Qt's way is very hard to maintain and read for this project.