Overhaul core library design
 * Renamed QDiscordXComponent classes to QDiscordX. They were too
verbose.
 * Modified the QDiscord class' interface to handle the new component
classes.
 * Removed the signal linkage between the component classes. They're now
linked by pointers to each other.
 * Added callbacks to most methods that used to just use signals. The
callbacks will let you respond to a specific event happening, instead of
listening in on all events and trying to somehow find out what caused
the event.
 * Objects returned from QDiscordState and callbacks will have a pointer
to a QDiscordRest instance, allowing you to call more convenient
non-static methods to make a request.
 * QDiscordState inherits a pure virtual class named
QDiscordAbstractState. Inherit QDiscordAbstractState and supply it to
QDiscordWs if you wish to implement your own payload handling. Inherit
QDiscordState and override a certain method if you want to use the
library's payload handling but modify a specific one.
 * Moved REST request building outside of QDiscordRest. They're not
static methods in their corresponding classes. For example,
QDiscordMessage::create. QDiscordRest has QDiscordRest::request, which
can be called to make a request with the stored auth token and with
automatic ratelimiting.
 * QDiscordRest now automatically handles ratelimiting. Any requests
that exceed the rate limits will be queued until they can be
successfully sent. The order in which requests arrive to Discord is not
guaranteed to be the same as the order in which the request was started
from the code, as multiple requests can happen in parallel.
 * Added some groundwork for the ability to use ETF with the WebSocket.
It still needs implementing.
 * Added an update method to QDiscordGuild.
 * Made the QDiscordID constructor explicit to eliminate some ambiguity
in methods.
 * Added the QDiscordMessage::TTS class to remove some ambiguity in
methods. You shouldn't need to explicitly use it, as it's implicitly
constructed from a bool and implicitly converts back to it.
 * Added a QDiscordUserAgent class to allow custom user agents per
QDiscordRest instance.
 * Added headers to easily include QDiscordRest, QDiscordState or
QDiscordWs, without including the other headers.